metascan_api/requirements.txt
fastapi
uvicorn
mysql-connector-python
pymysql
